Tests/LibWeb/Text/expected/link-element-rel-preload-load-event.txt

Got error event
Got load event
